src/main/java/com/switchfully/eurder/item/service/ItemService.java
package com.switchfully.eurder.item.service;

import com.switchfully.eurder.item.domain.Item;
import com.switchfully.eurder.item.domain.ItemRepository;
import com.switchfully.eurder.utils.Utils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class ItemService {

    private final ItemMapper itemMapper;
    private final ItemRepository itemRepository;

    @Autowired
    public ItemService(ItemMapper itemMapper, ItemRepository itemRepository) {
        this.itemMapper = itemMapper;
        this.itemRepository = itemRepository;
    }

    public ItemDto save(CreateItemDto createItemDto, String adminId) {
        Utils.adminAccess(adminId);

        Item item = itemMapper.fromDto(createItemDto);
        Item savedItem = itemRepository.save(item);
        return itemMapper.toDto(savedItem);
    }
}
